Image/ecommerce_image_processing/pyproject.toml
[project]
name = "extract-product-for-ecommerce"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-cloud-aiplatform>=1.108.0",
    "google-cloud-storage>=2.19.0",
    "google-genai>=1.29.0",
    "ipycanvas>=0.14.1",
    "ipykernel>=6.30.1",
    "ipython>=9.4.0",
    "ipywidgets>=8.1.7",
    "matplotlib>=3.10.5",
]